+++ WildAnalyticsSDK/WildAnalyticsSDK/Sources/WBAnalytics/WBAnalytics.swift
@@ -4,7 +4,7 @@ import Foundation
 import UIKit
 
 /// Public delegate of WBAnalytics
-public protocol WBAnalyticsDelegateProtocol: AnyObject {
+public protocol WildAnalyticsDelegateProtocol: AnyObject {
 
     /// Called when WB Tracker found an attributed deeplink that can be handled by the client
     /// - Parameter link: URL
@@ -48,14 +48,14 @@ public class WBAnalytics {
 
     lazy var configProvider: ConfigProvider = ConfigProvider(logger: logger)
 
-    private lazy var attributionTracker: WBTracker = {
-        WBTracker(apiKey: apiKey, logger: logger)
+    private lazy var attributionTracker: WildTracker = {
+        WildTracker(apiKey: apiKey, logger: logger)
     }()
 
     private let apiKey: String
     private let analyticsURL: URL
     private let interceptor: RequestInterceptor
-    private weak var delegate: WBAnalyticsDelegateProtocol?
+    private weak var delegate: WildAnalyticsDelegateProtocol?
 
     static var loggingOptions: LoggingOptions = .default
 
@@ -82,7 +82,7 @@ public class WBAnalytics {
         analyticsURL: URL,
         interceptor: RequestInterceptor,
         loggingOptions: LoggingOptions = .default,
-        delegate: WBAnalyticsDelegateProtocol? = nil
+        delegate: WildAnalyticsDelegateProtocol? = nil
     ) -> WBAnalytics {
         let analytics = WBAnalytics(
             apiKey: apiKey,